add all auth routes to one file i.e. auth.py
