+++ app.py
@@ -2,10 +2,7 @@ from fastapi import FastAPI, Depends
 from contextlib import asynccontextmanager
 from auth.jwt_bearer import JWTBearer
 from config.config import initiate_database
-from routes.login import router as LoginRouter
-from routes.auth.forgotPassword import router as ForgotPasswordRouter
-from routes.auth.resetPassword import router as ResetPasswordRouter 
-from routes.auth.login import router as LoginRouter
+from routes.auth import router as authRouter
 from routes.admin import router as AdminRouter
 from routes.admin_hr import router as AdminHRRouter
 from routes.employee import router as EmployeeRouter
@@ -35,9 +32,7 @@ async def read_root() -> dict:
 
 
 # Including routers
-app.include_router(LoginRouter,prefix="/auth/login", tags=["Login"])
-app.include_router(ForgotPasswordRouter, prefix="/auth/forgot-password", tags=["Forgot Password"])
-app.include_router(ResetPasswordRouter, prefix="/auth/reset-password", tags=["Reset Password"])
+app.include_router(authRouter,prefix="/auth", tags=["auth"])
 
 app.include_router(AdminRouter, tags=["Admin"], prefix="/admin")
 app.include_router(AdminHRRouter, tags=["Admin-HR"], prefix="/admin-hr")
